+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Always validate activity audience.
+
 ### Fixed
 
 - Fixed display of ">" in code blocks ([#134](https://codeberg.org/silverpill/mitra/pulls/134)).